Move FindWorkspaceRoot to sdk/nodejs/npm
We want to use FindWorkspaceRoot inside program test later on.
