+++ sdk/nodejs/npm/testdata/nested/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "nested-project",
+    "version": "1.0.0"
+}
