+++ sdk/nodejs/npm/testdata/workspace/packages/some-package/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "some-package",
+  "version": "1.0.0"
+}
