+++ sdk/nodejs/npm/testdata/workspace/project/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "project",
+    "version": "1.0.0"
+}
